dashboard/frontend/requirements.txt
streamlit
requests # Para fazer chamadas HTTP ao Flask
pandas
plotly-express